pyproject.toml
[tool.poetry]
name = "cashcow"
version = "0.1.0"
description = "Cash flow modeling system for rocket engine company"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "cashcow", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = {extras = ["email"], version = "^2.5.0"}
pyyaml = "^6.0.1"
click = "^8.1.7"
pandas = "^2.2.0"
matplotlib = "^3.8.0"
sqlalchemy = "^2.0.0"
aiosqlite = "^0.20.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
websockets = "^12.0"
aiofiles = "^24.1.0"
openpyxl = "^3.1.0"
reportlab = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.25.0"
pytest-mock = "^3.12.0"
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
psutil = "^7.0.0"
memory-profiler = "^0.61.0"
pytest-json-report = "^1.5.0"
rich = "^14.0.0"
numpy = ">=1.21.0,<2.0.0"
watchdog = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.poetry.scripts]
cashcow = "cashcow.cli.main:cli"